pyproject.toml
[project]
name = "ai-document-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.57.0",
]
